refactor(runtime): extract response interpretation into agent/turn — #7415
Move build_native_assistant_history and resolve_display_text into
turn/parse_response.rs, and extract the chat_result Ok-arm (observer
LlmResponse + cost recording, think-tag strip, native and text-fallback
tool-call parsing filtered to known names, parse-issue detection,
assistant-history construction) as interpret_chat_response(..) ->
InterpretedResponse. The response is taken by value; no ctx borrows are
held across the call. The protocol_suppressed / response_streamed_live
tuple elements remain pass-throughs of the call_provider outcome locals
in the loop.

The tool-specs glue now keeps the IterationToolSpecs struct alive
(ref-destructuring the fields the loop still uses inline) so it can be
passed to interpret_chat_response per the RUN_SHEET signature; the
known_tool_names binding is dropped (its only user moved). Re-export
hygiene per the S2 rule: build_native_assistant_history and the
loop_.rs detect_* re-exports are test-only/removed now that their lib
users live inside turn/.
